move modern CheckBox
