+++ src/components/core/CheckBox/CheckBox.stories.js
