+++ src/components/core/CheckBox/CheckBox.test.js
