+++ src/components/legacy/Checkbox/Checkbox.js
@@ -87,7 +87,7 @@ const CheckboxWrapper = styled.View`
 /**
  * @deprecated This compontent is considered legacy and should not be used in new code
  *
- * Use: components/modern/CheckBox instead
+ * Use: components/core/CheckBox instead
  */
 const Checkbox = (props: Props) => {
   const {
