+++ src/screens/WalletMigrationArchanova/SelectAssets/AssetList.js
@@ -26,7 +26,7 @@ import { useTranslation } from 'translations/translate';
 
 // Components
 import AssetListItem from 'components/modern/AssetListItem';
-import CheckBox from 'components/modern/CheckBox';
+import CheckBox from 'components/core/CheckBox';
 import Text from 'components/core/Text';
 
 // Constants
